Add option to control rerolling condition of parts
Previously, parts inside of a gun would have their condition rolled the
first time you interact with the gun. Now, this option can be used to
control this behavior.

Enabled: Parts will always be rerolled when interacting with a gun for
the first time. This will cause the apparent condition of the gun to
change. This was the previous behavior.

Disabled: Parts will not be rerolled when interacting with a gun for the
first time. The condition of the internal parts will inherit the
condition of the gun. This means the apparent condition of the gun will
not change. This is the new default behavior.